Add logout and register option

+++ classroomApp/views.py
@@ -5,16 +5,28 @@ from .forms import RoomForm
 from django.db.models import Q # type: ignore
 from django.contrib.auth.models import User # type: ignore
 from django.contrib.auth import authenticate, login, logout # type: ignore
+from django.contrib.auth.decorators import login_required # type: ignore
+from django.http import HttpResponse # type: ignore
 
 # rooms = [
 #     {'id': 1, 'name': 'Room 1'},
 #     {'id': 2, 'name': 'Room 2'},
 #     {'id': 3, 'name': 'Room 3'},
 # ]
+
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect("home")
+    
     if request.method == "POST":
         username = request.POST.get('username')
         password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "User does not exist")
 
         user = authenticate(request, username=username, password=password)
         if user is not None:
@@ -24,7 +36,35 @@ def loginPage(request):
         else:
             messages.error(request, "Invalid credentials")
 
-    context = {}
+    context = {'page': page}
+    return render(request, "base/login_register.html", context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect("login")
+
+def registerPage(request):
+    page = 'register'
+    if request.user.is_authenticated:
+            return redirect("home")
+    
+    if request.method == "POST":    
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+
+        if User.objects.filter(username=username).exists():
+            messages.error(request, "Account already exists, try logging in")
+            return redirect("login")
+        else:
+            try:
+                user = User.objects.create_user(username=username, password=password)
+                user.save()
+                messages.success(request, "User created successfully")
+                return redirect("login")
+            except:
+                messages.error(request, "An error occurred while creating user")
+
+    context = {'page': page}
     return render(request, "base/login_register.html", context)
 
 def home(request):
@@ -48,6 +88,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, "base/room.html", context)
 
+@login_required(login_url='/login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -58,9 +99,14 @@ def createRoom(request):
         
     return render(request, "base/room_form.html", context={'form': form})
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to edit this room")
+
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -69,8 +115,13 @@ def updateRoom(request, pk):
         
     return render(request, "base/room_form.html", context={'form': form})
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed to delete this room")
+    
     if request.method == "POST":
         room.delete()
         return redirect("home")
